+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## UNRELEASED
+
+### Changed
+
+- Replace DB Checker with Ping Checker
+
+
 ## 0.4.0 - 2017-02-16
 
 ### Added